[FEM] Elmer writer: sort out flux equation
- sort out flux equation - next step to refactor writer.py
