+++ src/Mod/Fem/CMakeLists.txt
@@ -258,6 +258,7 @@ SET(FemSolverElmerEquations_SRCS
     femsolver/elmer/equations/equation.py
     femsolver/elmer/equations/flow.py
     femsolver/elmer/equations/flux.py
+    femsolver/elmer/equations/flux_writer.py
     femsolver/elmer/equations/heat.py
     femsolver/elmer/equations/heat_writer.py
     femsolver/elmer/equations/linear.py
